two things have been done here. 1-a middleware file that was named "generalMiddleware" but only contained one component for validation was renamed to be more clear to "validationMiddleware" and the files that use this middleware were updated accoridingly. At the same time, the logic for the availabilty has been written. It is now being tested and refactored.
